src/main/java/code/kliangh/graphql/resolver/UserQuery.java
package code.kliangh.graphql.resolver;

import code.kliangh.user.entity.User;
import code.kliangh.user.repository.UserRepository;
import com.coxautodev.graphql.tools.GraphQLQueryResolver;
import org.springframework.stereotype.Component;

import javax.persistence.EntityNotFoundException;
import java.util.List;

@Component
public class UserQuery implements GraphQLQueryResolver {

    private UserRepository userRepository;

    public UserQuery(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public List<User> findAllUsers() {
        return userRepository.findAll();
    }

    public User user(String uid) {
        return userRepository.findById(uid).orElseThrow(EntityNotFoundException::new);
    }
}
